requirements.txt
python-docx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
googletrans
